+++ CHANGELOG.md
@@ -13,6 +13,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Exposed `ensureAuthToken` function for making authenticated GraphQL calls directly from platforms
   (#576)
 - Added `ownerId` prop to `<manifold-data-provision-button>`
+- Added test for `<manifold-plan-selector regions="">` (#644)
 
 ### Removed